+++ src/lightkurve/__init__.py
@@ -34,6 +34,7 @@ log = logging.getLogger(__name__)
 log.addHandler(logging.StreamHandler())
 
 from .version import __version__
+from .units import *
 from .time import *
 from .lightcurve import *
 from .lightcurvefile import *